+++ app/controllers/IndividualUserProfileController.java
@@ -84,6 +84,10 @@ public class IndividualUserProfileController extends Controller {
 
 
     public static Result deleteProfile(String userId) {
+        IndividualUserProfile profileToDelete = dbIndividualProfileService.get(new ObjectId(userId));
+        if(profileToDelete == null){
+            return notFound("No existing profile for this user: " + userId);
+        }
         WriteResult writeResult = dbIndividualProfileService.deleteById(new ObjectId(userId));
         if (writeResult.getN() == 1) {
             return ok();
